requirements.txt
streamlit
langchain 
langchain-groq 
python-dotenv